fix(l10n): Move getDirection to localizationUtils
