+++ content-test/common/reducers/Intl.test.js
@@ -1,25 +1,9 @@
 const Intl = require("common/reducers/Intl");
-const {getDirection} = Intl;
-const {DEFAULT_LOCALE} = require("common/constants");
 
 describe("Intl reducer", () => {
   it("should have the correct initial state", () => {
     const state = Intl(undefined, {});
     assert.deepEqual(state, Intl.INITIAL_STATE);
-  });
-  describe("getDirection", () => {
-    it("should return ltr with default strings", () => {
-      const expectedDir = "ltr";
-      assert.equal(getDirection(DEFAULT_LOCALE), expectedDir);
-    });
-    it("should gracefully handle 3 letter code locales", () => {
-      const expectedDir = "ltr";
-      assert.equal(getDirection("aus-IL"), expectedDir);
-    });
-    it("should return rtl for rtl languages", () => {
-      const expectedDir = "rtl";
-      assert.equal(getDirection("he-IL"), expectedDir);
-    });
   });
   describe("LOCALE_UPDATED", () => {
     it("should not update if action.data is missing", () => {
